# Faster byte[] copy