This makes me sad that this is even possible to do (+4% RPS)...

Issue: "Array.Copy & Buffer.BlockCopy x2 to x3 slower < 1kB" https://github.com/dotnet/coreclr/issues/2430

[![Performance Graph](http://aoa.blob.core.windows.net/aspnet/VectorizedCopy888.png)](http://aoa.blob.core.windows.net/aspnet/VectorizedCopy.png)

Until that is fixed, there is this, +4% RPS

Before

> Running 30s test @ http://.../plaintext
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.31ms   17.04ms 430.36ms   96.48%
    Req/Sec    63.61k    11.37k  121.07k    89.55%
  60603911 requests in 30.04s, 7.45GB read
  Socket errors: connect 35, read 0, write 0, timeout 0
Requests/sec: 2017281.16
Transfer/sec:    253.95MB

> Running 30s test @ http://.../plaintext
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.86ms   18.76ms 319.53ms   96.10%
    Req/Sec    65.93k     7.22k  233.05k    91.79%
  61092441 requests in 30.10s, 7.51GB read
  Socket errors: connect 35, read 0, write 0, timeout 0
Requests/sec: 2029658.31
Transfer/sec:    255.50MB

> Running 30s test @ http://.../plaintext
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.76ms   18.81ms 338.18ms   96.30%
    Req/Sec    65.90k     6.02k  141.64k    88.76%
  61119465 requests in 30.10s, 7.51GB read
  Socket errors: connect 35, read 0, write 0, timeout 0
Requests/sec: 2030569.81
Transfer/sec:    255.62MB


After

> Running 30s test @ http://.../plaintext
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     6.54ms   16.04ms 421.70ms   97.95%
    Req/Sec    66.30k    13.82k  277.14k    91.29%
  63126220 requests in 30.10s, 7.76GB read
  Socket errors: connect 35, read 0, write 0, timeout 0
Requests/sec: 2097240.79
Transfer/sec:    264.01MB

> Running 30s test @ http://.../plaintext
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     6.39ms   15.65ms 315.77ms   98.11%
    Req/Sec    68.62k     7.74k  234.24k    87.15%
  63593990 requests in 30.10s, 7.82GB read
  Socket errors: connect 35, read 0, write 0, timeout 8
Requests/sec: 2112790.91
Transfer/sec:    265.97MB

> Running 30s test @ http://.../plaintext
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     6.25ms   15.05ms 332.33ms   98.25%
    Req/Sec    66.43k    13.67k  283.71k    92.85%
  63515232 requests in 30.10s, 7.81GB read
  Socket errors: connect 35, read 0, write 0, timeout 0
Requests/sec: 2110181.96
Transfer/sec:    265.64MB