@Anderman 2016-01-06
@benaadams Any updates on this. I think it is still possible to make it faster than `Buffer.MemoryCopy`. I can make a PR for this but that would be without `using IllyriadGames.ByteArrayExtensions` (-;